Fixed MBean interface and class names. Clearing.
(cherry picked from commit 916313ee39b92df63ec1a0bd686abf1bdd4d2fcf)
