--- core/mux/jar/src/main/resources/META-INF/jboss-deployment-structure.xml
@@ -1,12 +1,0 @@
-<jboss-deployment-structure>
-
-    <deployment>
-
-        <dependencies>
-            <module name="org.slf4j" />
-            <module name="org.mobicents.diameter" />
-        </dependencies>
-
-    </deployment>
-
-</jboss-deployment-structure>
