+++ core/mux/pom.xml
@@ -60,6 +60,14 @@
           <resume>false</resume>
         </configuration>
       </plugin>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-checkstyle-plugin</artifactId>
+        <version>2.17</version>
+        <configuration>
+          <skip>true</skip>
+        </configuration>
+      </plugin>
     </plugins>
   </build>
 
@@ -67,28 +75,14 @@
     <profile>
       <id>maven-release</id>
       <modules>
-        <module>sar-jboss-4</module>
         <module>sar-jboss-7</module>
-        <module>sar-jboss-5</module>
-      </modules>
-    </profile>
-    <profile>
-      <id>jboss4</id>
-      <modules>
-        <module>sar-jboss-4</module>
-      </modules>
-    </profile>
-    <profile>
-      <id>jboss5</id>
-      <activation>
-        <activeByDefault>true</activeByDefault>
-      </activation>
-      <modules>
-        <module>sar-jboss-5</module>
       </modules>
     </profile>
     <profile>
       <id>jboss7</id>
+      <activation>
+        <activeByDefault>true</activeByDefault>
+      </activation>
       <modules>
         <module>sar-jboss-7</module>
       </modules>
