--- core/mux/sar-jboss-7/src/main/resources/META-INF/jboss-deployment-structure.xml
@@ -1,10 +1,0 @@
-<jboss-deployment-structure>
-      <ear-subdeployments-isolated>true</ear-subdeployments-isolated>
-
-      <deployment>
-        <dependencies>
-          <module name="org.mobicents.diameter" />
-        </dependencies>
-      </deployment>
-
-</jboss-deployment-structure>
